vuln-agent/requirements.txt
groq>=0.8.0
openpyxl>=3.1.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
pytest>=8.0.0
